requirements.txt
langchain==0.3.9
langchain-openai==0.2.8
langchain-anthropic==0.1.1
langchain-ollama==0.2.2
playwright==1.44.0
playwright-stealth==1.0.6
markdownify==0.13.1
python-dotenv==1.0.1
colorama==0.4.6
tiktoken==0.8.0
langchain-text-splitters==0.3.2
apify-client==1.5.0
